chnage float to double

+++ src/Adafruit_MAX31865.cpp
@@ -116,7 +116,7 @@ void Adafruit_MAX31865::setWires(max31865_numwires_t wires ) {
 double  Adafruit_MAX31865::temperature(double RTDnominal, double refResistor) {
   // http://www.analog.com/media/en/technical-documentation/application-notes/AN709_0.pdf
 
-  float Z1, Z2, Z3, Z4, Rt, temp;
+  double Z1, Z2, Z3, Z4, Rt, temp;
 
   Rt = readRTD();
   Rt /= 32768;
